website: initial docs for this provider
Until now this provider was only documented indirectly via an example of
using `null_resource` with provisioners. It's documented here now largely
for the sake of completeness, but also so that its primary docs can live
in the null provider's own repository rather than in the core Terraform
repository.

This also addresses #10898.