Allows to store anything on null_resource
The null_resource can be great to store any value. This commit add the
bility to store arbitrary values without recreating the resources (in
the `values` section) and to store `inputs` that are stored within the
tfstate, accessible as `outputs` and updated only when the resource is
(re-)created.

This can solve use case in issue #5